# app/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Eventflow",
        "description": "Multi-tenant Group Travel & Event Booking Infrastructure. Digitizes MICE events and destination weddings from venue discovery to final rooming list.",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/v1/events/{event_id}/waitlist": {
            "post": {
                "tags": [
                    "Waitlist"
                ],
                "summary": "Add guest to waitlist",
                "description": "Manually add a guest to the waitlist from the planner dashboard.\n(Guests can also auto-add themselves via a public route in Phase 2E).",
                "operationId": "add_waitlist_route_api_v1_events__event_id__waitlist_post",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "event_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Event Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WaitlistCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WaitlistResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "Waitlist"
                ],
                "summary": "List waitlist for an event",
                "description": "Get paginated waitlist entries. Computes relative position for guests\nwith `status=waiting` without rewriting the database.",
                "operationId": "list_event_waitlist_api_v1_events__event_id__waitlist_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "event_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Event Id"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "default": 1,
                            "title": "Page"
                        }
                    },
                    {
                        "name": "page_size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "default": 20,
                            "title": "Page Size"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Status"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WaitlistListResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/waitlists/{waitlist_id}/status": {
            "put": {
                "tags": [
                    "Waitlist"
                ],
                "summary": "Manually update waitlist status",
                "description": "Allow planners to cancel or expire a waitlist entry.\nIf an 'offered' entry is cancelled, it immediately promotes the next waiting guest.",
                "operationId": "update_status_route_api_v1_waitlists__waitlist_id__status_put",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "waitlist_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Waitlist Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WaitlistActionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WaitlistResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/public/waitlists": {
            "post": {
                "tags": [
                    "Public Waitlist"
                ],
                "summary": "Guest auto-adds to waitlist via magic link",
                "description": "Public route for guests to join a waitlist using their booking token.\nValidates token and maps the action to the correct guest and event context.",
                "operationId": "public_add_waitlist_route_api_v1_public_waitlists_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PublicWaitlistCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WaitlistResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "PublicWaitlistCreate": {
                "properties": {
                    "guest_token": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Guest Token"
                    },
                    "room_block_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Room Block Id"
                    },
                    "room_type": {
                        "type": "string",
                        "maxLength": 100,
                        "title": "Room Type",
                        "examples": [
                            "standard"
                        ]
                    }
                },
                "type": "object",
                "required": [
                    "guest_token",
                    "room_block_id",
                    "room_type"
                ],
                "title": "PublicWaitlistCreate",
                "description": "Payload for guests joining waitlist via magic link token."
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    },
                    "input": {
                        "title": "Input"
                    },
                    "ctx": {
                        "type": "object",
                        "title": "Context"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
            "WaitlistActionRequest": {
                "properties": {
                    "status": {
                        "type": "string",
                        "pattern": "^(offered|expired|cancelled)$",
                        "title": "Status"
                    }
                },
                "type": "object",
                "required": [
                    "status"
                ],
                "title": "WaitlistActionRequest",
                "description": "Payload when a planner manually changes waitlist status."
            },
            "WaitlistCreate": {
                "properties": {
                    "guest_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Guest Id"
                    },
                    "room_block_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Room Block Id"
                    },
                    "room_type": {
                        "type": "string",
                        "maxLength": 100,
                        "title": "Room Type",
                        "examples": [
                            "standard"
                        ]
                    }
                },
                "type": "object",
                "required": [
                    "guest_id",
                    "room_block_id",
                    "room_type"
                ],
                "title": "WaitlistCreate",
                "description": "Payload to add a guest to the waitlist."
            },
            "WaitlistListResponse": {
                "properties": {
                    "items": {
                        "items": {
                            "$ref": "#/components/schemas/WaitlistResponse"
                        },
                        "type": "array",
                        "title": "Items"
                    },
                    "total": {
                        "type": "integer",
                        "title": "Total"
                    },
                    "page": {
                        "type": "integer",
                        "title": "Page"
                    },
                    "page_size": {
                        "type": "integer",
                        "title": "Page Size"
                    }
                },
                "type": "object",
                "required": [
                    "items",
                    "total",
                    "page",
                    "page_size"
                ],
                "title": "WaitlistListResponse",
                "description": "Paginated list of waitlist entries."
            },
            "WaitlistResponse": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "tenant_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Tenant Id"
                    },
                    "event_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Event Id"
                    },
                    "guest_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Guest Id"
                    },
                    "room_block_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Room Block Id"
                    },
                    "room_type": {
                        "type": "string",
                        "title": "Room Type"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "offer_expires_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Offer Expires At"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Updated At"
                    },
                    "position": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Position"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "tenant_id",
                    "event_id",
                    "guest_id",
                    "room_block_id",
                    "room_type",
                    "status",
                    "offer_expires_at",
                    "created_at",
                    "updated_at"
                ],
                "title": "WaitlistResponse",
                "description": "Data returned to the planner or guest."
            },
...
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}